requirements.txt
requests==2.31.0
openai==1.12.0
matplotlib==3.8.2
plotly==5.18.0
pandas==2.2.0
coinbase==2.1.0
schedule==1.2.1
python-dotenv==1.0.1
perplexity==0.3.0 